requirements.txt
gradescope-utils>=0.3.1
subprocess32
pyyaml
